+++ ouroboros-network/testlib/Test/Ouroboros/Network/Diffusion/Testnet/Cardano.hs
@@ -16,7 +16,8 @@
 
 module Test.Ouroboros.Network.Diffusion.Testnet.Cardano (tests) where
 
-import Control.Exception (AssertionFailed (..), catch, evaluate, fromException)
+import Control.Exception (AssertionFailed (..), catch, displayException,
+           evaluate, fromException)
 import Control.Monad.Class.MonadFork
 import Control.Monad.Class.MonadTest (exploreRaces)
 import Control.Monad.Class.MonadTime.SI (DiffTime, Time (Time), addTime,
@@ -24,16 +25,19 @@ import Control.Monad.Class.MonadTime.SI (DiffTime, Time (Time), addTime,
 import Control.Monad.IOSim
 
 import Data.Bifoldable (bifoldMap)
-import Data.Bifunctor (first)
+import Data.Bifunctor (bimap, first)
+import Data.Char (ord)
 import Data.Dynamic (fromDynamic)
-import Data.Foldable (fold)
+import Data.Foldable (fold, foldr')
 import Data.IP qualified as IP
+import Data.List (intercalate, sort)
 import Data.List qualified as List
 import Data.List.Trace qualified as Trace
 import Data.Map (Map)
 import Data.Map qualified as Map
 import Data.Maybe (catMaybes, fromJust, fromMaybe, isJust, mapMaybe)
 import Data.Monoid (Sum (..))
+import Data.Ratio (Ratio)
 import Data.Set (Set)
 import Data.Set qualified as Set
 import Data.Time (secondsToDiffTime)
@@ -49,22 +53,23 @@ import Network.Mux.Trace qualified as Mx
 import Cardano.Network.ConsensusMode
 import Cardano.Network.PeerSelection.Bootstrap (UseBootstrapPeers (..),
            requiresBootstrapPeers)
-import Cardano.Network.PeerSelection.PeerTrustable (PeerTrustable (..))
-import Cardano.Network.Types (LedgerStateJudgement, NumberOfBigLedgerPeers (..))
-
 import Cardano.Network.PeerSelection.ExtraRootPeers qualified as Cardano
 import Cardano.Network.PeerSelection.ExtraRootPeers qualified as Cardano.ExtraPeers
 import Cardano.Network.PeerSelection.Governor.PeerSelectionState qualified as Cardano
 import Cardano.Network.PeerSelection.Governor.PeerSelectionState qualified as Cardano.ExtraState
+import Cardano.Network.PeerSelection.PeerTrustable (PeerTrustable (..))
+import Cardano.Network.PeerSelection.PublicRootPeers qualified as PublicRootPeers
+import Cardano.Network.Types (LedgerStateJudgement, NumberOfBigLedgerPeers (..))
 
 import Ouroboros.Network.Block (BlockNo (..))
-import Ouroboros.Network.BlockFetch (PraosFetchMode (..),
+import Ouroboros.Network.BlockFetch (FetchMode (..), PraosFetchMode (..),
            TraceFetchClientState (..))
 import Ouroboros.Network.ConnectionHandler (ConnectionHandlerTrace (..))
 import Ouroboros.Network.ConnectionId
 import Ouroboros.Network.ConnectionManager.Core qualified as CM
 import Ouroboros.Network.ConnectionManager.State qualified as CM
 import Ouroboros.Network.ConnectionManager.Types
+import Ouroboros.Network.Diffusion.Policies qualified as Diffusion
 import Ouroboros.Network.ExitPolicy (RepromoteDelay (..))
 import Ouroboros.Network.InboundGovernor qualified as IG
 import Ouroboros.Network.Mock.ConcreteBlock (BlockHeader)
@@ -81,6 +86,14 @@ import Ouroboros.Network.PeerSelection.State.LocalRootPeers (HotValency (..),
 import Ouroboros.Network.PeerSelection.State.LocalRootPeers qualified as LocalRootPeers
 import Ouroboros.Network.PeerSharing (PeerSharingResult (..))
 import Ouroboros.Network.Server qualified as Server
+import Ouroboros.Network.TxSubmission.Inbound.V2.Policy
+           (defaultTxDecisionPolicy, txInflightMultiplicity)
+import Ouroboros.Network.TxSubmission.Inbound.V2.State (inflightTxs)
+import Ouroboros.Network.TxSubmission.Inbound.V2.Types (TraceTxLogic (..),
+           TraceTxSubmissionInbound (..))
+import Ouroboros.Network.TxSubmission.Outbound (TxSubmissionProtocolError (..))
+
+import Simulation.Network.Snocket (BearerInfo (..), noAttenuation)
 
 import Test.Ouroboros.Network.ConnectionManager.Timeouts
 import Test.Ouroboros.Network.ConnectionManager.Utils
@@ -93,15 +106,16 @@ import Test.Ouroboros.Network.Diffusion.Node.Kernel
 import Test.Ouroboros.Network.Diffusion.Testnet.Cardano.Simulation
 import Test.Ouroboros.Network.InboundGovernor.Utils
 import Test.Ouroboros.Network.LedgerPeers (LedgerPools (..))
+import Test.Ouroboros.Network.TxSubmission.TxLogic (ArbTxDecisionPolicy (..))
+import Test.Ouroboros.Network.TxSubmission.Types (Tx (..), TxId)
 import Test.Ouroboros.Network.Utils hiding (SmallDelay, debugTracer)
 
-import Simulation.Network.Snocket (BearerInfo (..))
 
-import Cardano.Network.PeerSelection.PublicRootPeers qualified as PublicRootPeers
 import Test.QuickCheck
 import Test.QuickCheck.Monoids
 import Test.Tasty
 import Test.Tasty.QuickCheck (testProperty)
+
 
 tests :: TestTree
 tests =
@@ -157,6 +171,10 @@ tests =
                                  prop_only_bootstrap_peers_in_fallback_state_iosimpor
     , nightlyTest $ testProperty "no non trustable peers before caught up state"
                                  prop_no_non_trustable_peers_before_caught_up_state_iosimpor
+    , testGroup "Tx Submission"
+      [ nightlyTest $ testProperty "no protocol errors"
+                                   prop_no_txSubmission_error_iosimpor
+      ]
     , testGroup "Churn"
       [ nightlyTest $ testProperty "no timeouts"
                                    prop_churn_notimeouts_iosimpor
@@ -238,6 +256,14 @@ tests =
                      unit_peer_sharing
       , testProperty "don't peershare the unwilling"
                       prop_no_peershare_unwilling_iosim
+      ]
+    , testGroup "Tx Submission"
+      [ testProperty "no protocol errors"
+                     prop_no_txSubmission_error_iosim
+      , testProperty "all transactions"
+                     unit_txSubmission_allTransactions
+      , testProperty "inflight coverage"
+                     prop_check_inflight_ratio
       ]
     , testGroup "Churn"
       [ testProperty "no timeouts" prop_churn_notimeouts_iosim
@@ -346,7 +372,7 @@ unit_cm_valid_transitions =
             , abiSDUSize                 = LargeSDU
             }
       ds = DiffusionScript
-            (SimArgs 1 10)
+            (SimArgs 1 10 defaultTxDecisionPolicy)
             (Script ((Map.empty, ShortDelay) :| [(Map.empty, LongDelay)]))
             [ ( NodeArgs
                   (-2)
@@ -388,7 +414,8 @@ unit_cm_valid_transitions =
                     [DNSLookupDelay {getDNSLookupDelay = 0.072}]))
                   Nothing
                   False
-                  (Script (FetchModeBulkSync :| [FetchModeBulkSync]))
+                  (Script (PraosFetchMode FetchModeBulkSync :| [PraosFetchMode FetchModeBulkSync]))
+                  []
                   , [JoinNetwork 0.5]
                 )
               , ( NodeArgs
@@ -429,7 +456,8 @@ unit_cm_valid_transitions =
                   (Script (DNSLookupDelay {getDNSLookupDelay = 0.125} :| []))
                   (Just (BlockNo 2))
                   False
-                  (Script (FetchModeDeadline :| []))
+                  (Script (PraosFetchMode FetchModeDeadline :| []))
+                  []
                   , [JoinNetwork 1.484_848_484_848]
                 )
             ]
@@ -564,7 +592,7 @@ unit_connection_manager_trace_coverage =
 
     script@(DiffusionScript _ _ nodes) =
       DiffusionScript
-        (SimArgs 1 20)
+        (SimArgs 1 20 defaultTxDecisionPolicy)
         (singletonTimedScript Map.empty)
         [ -- a relay node
           (NodeArgs {
@@ -592,7 +620,9 @@ unit_connection_manager_trace_coverage =
              naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0.1} :| []),
              naChainSyncExitOnBlockNo = Nothing,
              naChainSyncEarlyExit = False,
-             naFetchModeScript = Script (FetchModeDeadline :| [])
+             naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+             naTxs = []
+
            }
           , [JoinNetwork 0]
           )
@@ -626,7 +656,8 @@ unit_connection_manager_trace_coverage =
              naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0.1} :| []),
              naChainSyncExitOnBlockNo = Nothing,
              naChainSyncEarlyExit = False,
-             naFetchModeScript = Script (FetchModeDeadline :| [])
+             naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+             naTxs = []
            }
           , [JoinNetwork 0]
           )
@@ -688,7 +719,7 @@ unit_connection_manager_transitions_coverage =
 
     script@(DiffusionScript _ _ nodes) =
       DiffusionScript
-        (SimArgs 1 20)
+        (SimArgs 1 20 defaultTxDecisionPolicy)
         (singletonTimedScript Map.empty)
         [ -- a relay node
           (NodeArgs {
@@ -716,7 +747,8 @@ unit_connection_manager_transitions_coverage =
              naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0.1} :| []),
              naChainSyncExitOnBlockNo = Nothing,
              naChainSyncEarlyExit = False,
-             naFetchModeScript = Script (FetchModeDeadline :| [])
+             naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+             naTxs = []
            }
           , [JoinNetwork 0]
           )
@@ -750,7 +782,8 @@ unit_connection_manager_transitions_coverage =
              naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0.1} :| []),
              naChainSyncExitOnBlockNo = Nothing,
              naChainSyncEarlyExit = False,
-             naFetchModeScript = Script (FetchModeDeadline :| [])
+             naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+             naTxs = []
            }
           , [JoinNetwork 0]
           )
@@ -786,6 +819,261 @@ prop_inbound_governor_trace_coverage defaultBearerInfo diffScript =
    -- TODO: Add checkCoverage here
    in tabulate "inbound governor trace" eventsSeenNames
       True
+
+-- | This test check that we don't have any tx submission protocol error
+--
+prop_no_txSubmission_error :: SimTrace Void
+                           -> Int
+                           -> Property
+prop_no_txSubmission_error ioSimTrace traceNumber =
+  let events = Trace.toList
+             . fmap (\(WithTime t (WithName _ b)) -> (t, b))
+             . withTimeNameTraceEvents
+                @DiffusionTestTrace
+                @NtNAddr
+             . Trace.take traceNumber
+             $ ioSimTrace
+
+   in counterexample (intercalate "\n" $ map show $ events)
+    $ all (\case
+             (_, DiffusionInboundGovernorTrace (IG.TrMuxErrored _ err)) ->
+               case fromException err of
+                 Just ProtocolErrorRequestBlocking               -> False
+                 Just ProtocolErrorRequestedNothing              -> False
+                 Just ProtocolErrorAckedTooManyTxids             -> False
+                 Just (ProtocolErrorRequestedTooManyTxids _ _ _) -> False
+                 Just ProtocolErrorRequestNonBlocking            -> False
+                 Just ProtocolErrorRequestedUnavailableTx        -> False
+                 _                                               -> True
+             _                                                   -> True
+          )
+          events
+
+prop_no_txSubmission_error_iosimpor
+  :: AbsBearerInfo -> DiffusionScript -> Property
+prop_no_txSubmission_error_iosimpor
+  = testWithIOSimPOR prop_no_txSubmission_error short_trace
+
+prop_no_txSubmission_error_iosim
+  :: AbsBearerInfo -> DiffusionScript -> Property
+prop_no_txSubmission_error_iosim
+  = testWithIOSim prop_no_txSubmission_error long_trace
+
+
+-- | This test checks that even in a scenario where nodes keep disconnecting,
+-- but eventually stay online. We manage to get all transactions.
+--
+unit_txSubmission_allTransactions :: ArbTxDecisionPolicy
+                                  -> NonEmptyList (Tx TxId)
+                                  -> NonEmptyList (Tx TxId)
+                                  -> Property
+unit_txSubmission_allTransactions (ArbTxDecisionPolicy decisionPolicy)
+                                  (NonEmpty txsA)
+                                  (NonEmpty txsB) =
+  let localRootConfig = LocalRootConfig
+                          DoNotAdvertisePeer
+                          InitiatorAndResponderDiffusionMode
+                          IsNotTrustable
+      diffScript =
+        DiffusionScript
+          (SimArgs 1 10 decisionPolicy)
+          (singletonTimedScript Map.empty)
+          [(NodeArgs
+              (-3)
+              InitiatorAndResponderDiffusionMode
+              (Just 224)
+              Map.empty
+              PraosMode
+              (Script (DontUseBootstrapPeers :| []))
+              (TestAddress (IPAddr (read "0.0.0.0") 0))
+              PeerSharingDisabled
+              [(2,2,Map.fromList [(RelayAccessAddress "0.0.0.1" 0, localRootConfig)])]
+              (Script (LedgerPools [] :| []))
+              (let targets =
+                    PeerSelectionTargets {
+                      targetNumberOfRootPeers = 1,
+                      targetNumberOfKnownPeers = 1,
+                      targetNumberOfEstablishedPeers = 1,
+                      targetNumberOfActivePeers = 1,
+
+                      targetNumberOfKnownBigLedgerPeers = 0,
+                      targetNumberOfEstablishedBigLedgerPeers = 0,
+                      targetNumberOfActiveBigLedgerPeers = 0
+                    }
+               in (targets, targets))
+              (Script (DNSTimeout {getDNSTimeout = 10} :| []))
+              (Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []))
+              Nothing
+              False
+              (Script (PraosFetchMode FetchModeDeadline :| []))
+              uniqueTxsA
+          , [JoinNetwork 0])
+          , (NodeArgs
+               (-1)
+               InitiatorAndResponderDiffusionMode
+               (Just 2)
+               Map.empty
+               PraosMode
+               (Script (DontUseBootstrapPeers :| []))
+               (TestAddress (IPAddr (read "0.0.0.1") 0))
+               PeerSharingDisabled
+               [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.0" 0, localRootConfig)])]
+               (Script (LedgerPools [] :| []))
+               (let targets =
+                      PeerSelectionTargets {
+                        targetNumberOfRootPeers = 1,
+                        targetNumberOfKnownPeers = 1,
+                        targetNumberOfEstablishedPeers = 1,
+                        targetNumberOfActivePeers = 1,
+
+                        targetNumberOfKnownBigLedgerPeers = 0,
+                        targetNumberOfEstablishedBigLedgerPeers = 0,
+                        targetNumberOfActiveBigLedgerPeers = 0
+                      }
+                in (targets, targets)
+               )
+               (Script (DNSTimeout {getDNSTimeout = 10} :| [ ]))
+               (Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []))
+               Nothing
+               False
+               (Script (PraosFetchMode FetchModeDeadline :| []))
+               uniqueTxsB
+         , [JoinNetwork 0])
+         ]
+   in checkAllTransactions (runSimTrace
+                              (diffusionSimulation noAttenuation
+                                                   diffScript
+                                                   iosimTracer)
+                           )
+                           500_000 -- ^ Running for 500k might not be enough.
+  where
+    -- We need to make sure the transactions are unique, this simplifies
+    -- things.
+    --
+    -- TODO: the generator ought to give us unique `TxId`s.
+    uniqueTxsA = map (\(t, i) -> t { getTxId = List.foldl' (+) 0 (map ord "0.0.0.0") + i })
+                     (zip txsA [0 :: TxId ..])
+    uniqueTxsB = map (\(t, i) -> t { getTxId = List.foldl' (+) 0 (map ord "0.0.0.1") + i })
+                     (zip txsB [100 :: TxId ..])
+
+    -- This checks the property that after running the simulation for a while
+    -- both nodes manage to get all valid transactions.
+    --
+    checkAllTransactions :: SimTrace Void
+                         -> Int
+                         -> Property
+    checkAllTransactions ioSimTrace traceNumber =
+      let trace  = Trace.take traceNumber ioSimTrace
+
+          events = fmap (\(WithTime t (WithName name b)) -> WithName name (WithTime t b))
+                 . withTimeNameTraceEvents
+                    @DiffusionTestTrace
+                    @NtNAddr
+                 $ trace
+
+          -- Build the accepted (sorted) txids map for each peer
+          --
+          sortedAcceptedTxidsMap :: Map NtNAddr [TxId]
+          sortedAcceptedTxidsMap =
+              foldr (\l r ->
+                List.foldl' (\rr (WithName n (WithTime _ x)) ->
+                  case x of
+                    -- When we add txids to the mempool, we collect them
+                    -- into the map
+                    DiffusionTxSubmissionInbound (TraceTxInboundAddedToMempool txids _) ->
+                      Map.alter (maybe (Just txids) (Just . sort . (txids ++))) n rr
+                    -- if a node would be killed, we could download some txs
+                    -- multiple times, but this is not possible in the schedule
+                    _ -> rr) r l
+              ) Map.empty
+            . Trace.toList
+            . splitWithNameTrace
+            $ events
+
+          -- Construct the list of valid (sorted) txs from peer A and peer B.
+          -- This is essentially our goal lists
+          --
+          (validSortedTxidsA, validSortedTxidsB) =
+            let f = sort
+                  . map (\Tx {getTxId} -> getTxId)
+                  . filter (\Tx {getTxValid} -> getTxValid)
+             in bimap f f (uniqueTxsA, uniqueTxsB)
+
+      in -- counterexample (intercalate "\n" $ map show $ Trace.toList events)
+          counterexample (Trace.ppTrace show (ppSimEvent 0 0 0) trace)
+        $ counterexample ("accepted txids map: " ++ show sortedAcceptedTxidsMap)
+        $ counterexample ("A: unique txs: " ++ show uniqueTxsA)
+        $ counterexample ("A: valid transactions that should be accepted: " ++ show validSortedTxidsA)
+        $ counterexample ("B: unique txs: " ++ show uniqueTxsB)
+        $ counterexample ("B: valid transactions that should be accepted: " ++ show validSortedTxidsB)
+        $ label ("number of valid tx transferred: " ++ renderRanges 10 (getSum . foldMap (Sum . List.length) $ sortedAcceptedTxidsMap))
+
+         -- Success criteria, after running for 500k events, we check the map
+         -- for the two nodes involved in the simulation and verify that indeed
+         -- each peer managed to learn about the other peer' transactions.
+         --
+        $ case Map.lookup (TestAddress (IPAddr (read "0.0.0.0") 0)) sortedAcceptedTxidsMap
+               of
+           Just acceptedTxidsA ->
+             counterexample "0.0.0.0" $
+                  acceptedTxidsA === validSortedTxidsB
+           Nothing | [] <- validSortedTxidsB -> property True
+                   | otherwise -> counterexample "Didn't found any entry in the map!" False
+        .&&.
+          case Map.lookup (TestAddress (IPAddr (read "0.0.0.1") 0)) sortedAcceptedTxidsMap
+               of
+           Just acceptedTxidsB ->
+             counterexample "0.0.0.1" $
+                  acceptedTxidsB === validSortedTxidsA
+           Nothing | [] <- validSortedTxidsA -> property True
+                   | otherwise -> counterexample "Didn't found any entry in the map!" False
+
+
+-- | This test checks the ratio of the inflight txs against the allowed by the
+-- TxDecisionPolicy.
+--
+prop_check_inflight_ratio :: AbsBearerInfo
+                          -> DiffusionScript
+                          -> Property
+prop_check_inflight_ratio bi ds@(DiffusionScript simArgs _ _) =
+  let sim :: forall s . IOSim s Void
+      sim = diffusionSimulation (toBearerInfo bi)
+                                ds
+                                iosimTracer
+
+      events :: Events DiffusionTestTrace
+      events = Signal.eventsFromList
+             . Trace.toList
+             . fmap ( (\(WithTime t (WithName _ b)) -> (t, b))
+                    )
+             . withTimeNameTraceEvents
+                @DiffusionTestTrace
+                @NtNAddr
+             . Trace.take 500000
+             $ runSimTrace
+             $ sim
+
+      inflightTxsMap =
+          foldr'
+            (\(_, m) r -> Map.unionWith (max) m r
+            )
+            Map.empty
+        $ Signal.eventsToList
+        $ Signal.selectEvents
+           (\case
+               DiffusionTxLogic (TraceSharedTxState _ d) -> Just (inflightTxs d)
+               _                                         -> Nothing
+           )
+        $ events
+
+      txDecisionPolicy = saTxDecisionPolicy simArgs
+
+   in tabulate "Max observeed ratio of inflight multiplicity by the max stipulated by the policy"
+               (map (\m -> "has " ++ show m ++ " in flight - ratio: "
+                    ++ show @(Ratio Int) (fromIntegral m / fromIntegral (txInflightMultiplicity txDecisionPolicy))
+                    )
+                    (Map.elems inflightTxsMap))
+    $ True
 
 -- | This test coverage of InboundGovernor transitions.
 --
@@ -909,10 +1197,10 @@ prop_only_bootstrap_peers_in_fallback_state ioSimTrace traceNumber =
 
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
-              Signal.fromChangeEvents Killed -- Default to TrKillingNode
+              Signal.fromChangeEvents Terminated -- Default to TrKillingNode
             . Signal.selectEvents
                 (\case TrJoiningNetwork -> Just Joined
-                       TrKillingNode    -> Just Killed
+                       TrTerminated     -> Just Terminated
                        _                -> Nothing
                 )
             . selectDiffusionSimulationTrace
@@ -938,10 +1226,10 @@ prop_only_bootstrap_peers_in_fallback_state ioSimTrace traceNumber =
           trIsNodeAlive :: Signal Bool
           trIsNodeAlive =
                 not . Set.null
-            <$> Signal.keyedUntil (fromJoinedOrKilled (Set.singleton ())
-                                                      Set.empty)
-                                  (fromJoinedOrKilled Set.empty
-                                                      (Set.singleton ()))
+            <$> Signal.keyedUntil (fromJoinedOrTerminated (Set.singleton ())
+                                                           Set.empty)
+                                  (fromJoinedOrTerminated  Set.empty
+                                                          (Set.singleton ()))
                                   (const False)
                                   trJoinKillSig
 
@@ -952,8 +1240,14 @@ prop_only_bootstrap_peers_in_fallback_state ioSimTrace traceNumber =
               -- frequently and disconnection timeouts we have to increase
               -- this value
               300 -- seconds
-              (\(knownPeers, useBootstrapPeers, trustedPeers, lsj, isAlive) ->
-                if isAlive && requiresBootstrapPeers useBootstrapPeers lsj
+              (\( knownPeers
+                , useBootstrapPeers
+                , trustedPeers
+                , ledgerStateJudgement
+                , isAlive
+                ) ->
+                if    isAlive
+                   && requiresBootstrapPeers useBootstrapPeers ledgerStateJudgement
                    then knownPeers `Set.difference` trustedPeers
                    else Set.empty
               )
@@ -961,7 +1255,7 @@ prop_only_bootstrap_peers_in_fallback_state ioSimTrace traceNumber =
                       <*> govUseBootstrapPeers
                       <*> govTrustedPeers
                       <*> govLedgerStateJudgement
-                       <*> trIsNodeAlive
+                      <*> trIsNodeAlive
               )
        in counterexample (List.intercalate "\n" $ map show $ Signal.eventsToList events)
         $ signalProperty 20 show
@@ -981,6 +1275,9 @@ prop_only_bootstrap_peers_in_fallback_state_iosim
 
 -- | Same as PeerSelection test 'prop_governor_no_non_trustable_peers_before_caught_up_state'
 --
+-- NOTE: the only difference from `prop_only_bootstrap_peers_in_fallback_state`
+-- is that the test verifies we only have bootstrap peers after
+-- `UseBootstrapPeers` was set.  We use slightly lower timeout.
 prop_no_non_trustable_peers_before_caught_up_state :: SimTrace Void
                                                    -> Int
                                                    -> Property
@@ -1017,7 +1314,7 @@ prop_no_non_trustable_peers_before_caught_up_state ioSimTrace traceNumber =
             selectDiffusionPeerSelectionState
               (Cardano.bootstrapPeersFlag . Governor.extraState)
               events
-
+      
           govLedgerStateJudgement :: Signal LedgerStateJudgement
           govLedgerStateJudgement =
             selectDiffusionPeerSelectionState (Cardano.ledgerStateJudgement . Governor.extraState)
@@ -1036,18 +1333,18 @@ prop_no_non_trustable_peers_before_caught_up_state ioSimTrace traceNumber =
               )
               events
 
-          govHasOnlyBootstrapPeers :: Signal Bool
-          govHasOnlyBootstrapPeers =
-            selectDiffusionPeerSelectionState
-              (Cardano.hasOnlyBootstrapPeers . Governor.extraState)
-              events
+          -- govHasOnlyBootstrapPeers :: Signal Bool
+          -- govHasOnlyBootstrapPeers =
+          --   selectDiffusionPeerSelectionState
+          --     (Cardano.hasOnlyBootstrapPeers . Governor.extraState)
+          --     events
 
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
-              Signal.fromChangeEvents Killed -- Default to TrKillingNode
+              Signal.fromChangeEvents Terminated -- Default to TrKillingNode
             . Signal.selectEvents
                 (\case TrJoiningNetwork -> Just Joined
-                       TrKillingNode    -> Just Killed
+                       TrTerminated     -> Just Terminated
                        _                -> Nothing
                 )
             . selectDiffusionSimulationTrace
@@ -1062,28 +1359,39 @@ prop_no_non_trustable_peers_before_caught_up_state ioSimTrace traceNumber =
           trIsNodeAlive :: Signal Bool
           trIsNodeAlive =
                 not . Set.null
-            <$> Signal.keyedUntil (fromJoinedOrKilled (Set.singleton ())
-                                                      Set.empty)
-                                  (fromJoinedOrKilled Set.empty
-                                                      (Set.singleton ()))
+            <$> Signal.keyedUntil (fromJoinedOrTerminated (Set.singleton ())
+                                                           Set.empty)
+                                  (fromJoinedOrTerminated  Set.empty
+                                                          (Set.singleton ()))
                                   (const False)
                                   trJoinKillSig
 
           keepNonTrustablePeersTooLong :: Signal (Set NtNAddr)
           keepNonTrustablePeersTooLong =
             Signal.keyedTimeout
-              10 -- seconds
-              (\( knownPeers, trustedPeers
-                , useBootstrapPeers, lsj, hasOnlyBootstrapPeers, isAlive) ->
-                  if isAlive && hasOnlyBootstrapPeers && requiresBootstrapPeers useBootstrapPeers lsj
+              -- it might take `closeConnectionTimeout` (120)s to close a connection.
+              Diffusion.closeConnectionTimeout
+              -- timeout arming function
+              (\( knownPeers
+                , trustedPeers
+                , useBootstrapPeers
+                , ledgerStateJudgement
+                , isAlive
+                ) ->
+                  -- arm the signal as soon as
+                  -- * the node is live
+                  -- * we are in a state which requires bootstrap peers
+                  -- * there are non trusted known peers
+                  if    isAlive
+                     && requiresBootstrapPeers useBootstrapPeers ledgerStateJudgement
                      then Set.difference knownPeers trustedPeers
                      else Set.empty
               )
-              ((,,,,,) <$> govKnownPeers
+              -- signal
+              ((,,,,) <$> govKnownPeers
                        <*> govTrustedPeers
                        <*> govUseBootstrapPeers
                        <*> govLedgerStateJudgement
-                       <*> govHasOnlyBootstrapPeers
                        <*> trIsNodeAlive
               )
 
@@ -1112,7 +1420,7 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
   where
     script :: DiffusionScript
     script =
-      DiffusionScript (SimArgs 1 10)
+      DiffusionScript (SimArgs 1 10 defaultTxDecisionPolicy)
         (singletonTimedScript Map.empty)
         [ ( NodeArgs (-6) InitiatorAndResponderDiffusionMode (Just 180)
               (Map.fromList [(RelayAccessDomain "test2" 65_535, DoAdvertisePeer)])
@@ -1136,7 +1444,8 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
               (Script (DNSLookupDelay {getDNSLookupDelay = 0.067} :| [DNSLookupDelay {getDNSLookupDelay = 0.097},DNSLookupDelay {getDNSLookupDelay = 0.101},DNSLookupDelay {getDNSLookupDelay = 0.096},DNSLookupDelay {getDNSLookupDelay = 0.051}]))
               Nothing
               False
-              (Script (FetchModeDeadline :| []))
+              (Script (PraosFetchMode FetchModeDeadline :| []))
+              []
           , [JoinNetwork 1.742_857_142_857
             ,Reconfigure 6.333_333_333_33 [(1,1,Map.fromList [(RelayAccessDomain "test2" 65_535,LocalRootConfig DoAdvertisePeer InitiatorAndResponderDiffusionMode IsNotTrustable)]),
                                         (1,1,Map.fromList [(RelayAccessAddress "0:6:0:3:0:6:0:5" 65_530,LocalRootConfig DoAdvertisePeer InitiatorAndResponderDiffusionMode IsNotTrustable)
@@ -1169,7 +1478,8 @@ unit_4177 = prop_inbound_governor_transitions_coverage absNoAttenuation script
                      ]))
              Nothing
              False
-             (Script (FetchModeDeadline :| []))
+             (Script (PraosFetchMode FetchModeDeadline :| []))
+             []
           , [JoinNetwork 0.183_783_783_783
             ,Reconfigure 4.533_333_333_333 [(1,1,Map.empty)]
             ]
@@ -1227,10 +1537,10 @@ prop_track_coolingToCold_demotions ioSimTracer traceNumber =
 
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
-              Signal.fromChangeEvents Killed -- Default to TrKillingNode
+              Signal.fromChangeEvents Terminated -- Default to TrKillingNode
             . Signal.selectEvents
                 (\case TrJoiningNetwork -> Just Joined
-                       TrKillingNode    -> Just Killed
+                       TrTerminated     -> Just Terminated
                        _                -> Nothing
                 )
             . selectDiffusionSimulationTrace
@@ -1245,10 +1555,10 @@ prop_track_coolingToCold_demotions ioSimTracer traceNumber =
           trIsNodeAlive :: Signal Bool
           trIsNodeAlive =
                 not . Set.null
-            <$> Signal.keyedUntil (fromJoinedOrKilled (Set.singleton ())
-                                                      Set.empty)
-                                  (fromJoinedOrKilled Set.empty
-                                                      (Set.singleton ()))
+            <$> Signal.keyedUntil (fromJoinedOrTerminated (Set.singleton ())
+                                                           Set.empty)
+                                  (fromJoinedOrTerminated  Set.empty
+                                                          (Set.singleton ()))
                                   (const False)
                                   trJoinKillSig
 
@@ -1709,7 +2019,7 @@ prop_diffusion_dns_can_recover ioSimTrace traceNumber =
                                   (recovered + 1)
                                   t
                                   evs
-        DiffusionDiffusionSimulationTrace TrReconfiguringNode ->
+        DiffusionSimulationTrace TrReconfiguringNode ->
           verify Map.empty ttlMap recovered t evs
         _ -> verify toRecover ttlMap recovered time evs
 
@@ -1754,7 +2064,7 @@ unit_4191 = testWithIOSim prop_diffusion_dns_can_recover long_trace absInfo scri
         }
     script =
       DiffusionScript
-        (SimArgs 1 20)
+        (SimArgs 1 20 defaultTxDecisionPolicy)
         (singletonTimedScript $
            Map.fromList
              [ (("test2", DNS.A), Left [ (read "810b:4c8a:b3b5:741:8c0c:b437:64cf:1bd9", 300)
@@ -1819,11 +2129,12 @@ unit_4191 = testWithIOSim prop_diffusion_dns_can_recover long_trace absInfo scri
                                                                    ]))
             Nothing
             False
-            (Script (FetchModeDeadline :| []))
+            (Script (PraosFetchMode FetchModeDeadline :| []))
+            []
             , [ JoinNetwork 6.710_144_927_536
               , Kill 7.454_545_454_545
               , JoinNetwork 10.763_157_894_736
-              , Reconfigure 0.415_384_615_384 [(1,1,Map.empty)
+              , Reconfigure 0.415_384_615_384 [(1,1,Map.fromList [])
               , (1,1,Map.empty)]
               , Reconfigure 15.550_561_797_752 [(1,1,Map.empty)
               , (1,1,Map.fromList [(RelayAccessDomain "test2" 15,LocalRootConfig DoAdvertisePeer InitiatorAndResponderDiffusionMode IsNotTrustable)])]
@@ -1881,7 +2192,7 @@ prop_connect_failure (AbsIOError ioerr) =
 
     script =
       DiffusionScript
-        (SimArgs 1 20)
+        (SimArgs 1 20 defaultTxDecisionPolicy)
         (singletonTimedScript Map.empty)
         [ (NodeArgs {
             naSeed = 0,
@@ -1908,7 +2219,8 @@ prop_connect_failure (AbsIOError ioerr) =
             naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []),
             naChainSyncExitOnBlockNo = Nothing,
             naChainSyncEarlyExit = False,
-            naFetchModeScript = Script (FetchModeDeadline :| [])
+            naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+            naTxs = []
           }
           , [JoinNetwork 10]
           ),
@@ -1937,7 +2249,8 @@ prop_connect_failure (AbsIOError ioerr) =
             naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []),
             naChainSyncExitOnBlockNo = Nothing,
             naChainSyncEarlyExit = False,
-            naFetchModeScript = Script (FetchModeDeadline :| [])
+            naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+            naTxs = []
           }
           , [JoinNetwork 0]
           )
@@ -1968,13 +2281,15 @@ prop_accept_failure (AbsIOError ioerr) =
            counterexample (show evs)
          . (if isFatalAccept ioerr
            then -- verify that the node was killed by the right exception
-                any (\case
+                counterexample ("fatal exception " ++ displayException ioerr ++ " not propagated")
+              . any (\case
                       TrErrored e | Just e' <- fromException e
                                   , e' == ioerr
                                   -> True
                       _           -> False)
            else -- verify that the node was not killed by the `ioerr` exception
-                all (\case
+                counterexample ("non-fatal exception " ++ displayException ioerr ++ " propagated")
+              . all (\case
                       TrErrored {} -> False
                       _            -> True)
            )
@@ -2007,7 +2322,7 @@ prop_accept_failure (AbsIOError ioerr) =
 
     script =
       DiffusionScript
-        (SimArgs 1 20)
+        (SimArgs 1 20 defaultTxDecisionPolicy)
         (singletonTimedScript Map.empty)
         [ (NodeArgs {
             naSeed = 0,
@@ -2034,7 +2349,8 @@ prop_accept_failure (AbsIOError ioerr) =
             naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []),
             naChainSyncExitOnBlockNo = Nothing,
             naChainSyncEarlyExit = False,
-            naFetchModeScript = Script (FetchModeDeadline :| [])
+            naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+            naTxs = []
           }
           , [JoinNetwork 10]
           ),
@@ -2063,7 +2379,8 @@ prop_accept_failure (AbsIOError ioerr) =
             naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []),
             naChainSyncExitOnBlockNo = Nothing,
             naChainSyncEarlyExit = False,
-            naFetchModeScript = Script (FetchModeDeadline :| [])
+            naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+            naTxs = []
           }
           , [JoinNetwork 0]
           )
@@ -2550,10 +2867,10 @@ prop_diffusion_target_established_local ioSimTrace traceNumber =
 
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
-              Signal.fromChangeEvents Killed -- Default to TrKillingNode
+              Signal.fromChangeEvents Terminated -- Default to TrKillingNode
             . Signal.selectEvents
                 (\case TrJoiningNetwork -> Just Joined
-                       TrKillingNode    -> Just Killed
+                       TrKillingNode    -> Just Terminated
                        _                -> Nothing
                 )
             . selectDiffusionSimulationTrace
@@ -2568,10 +2885,10 @@ prop_diffusion_target_established_local ioSimTrace traceNumber =
           trIsNodeAlive :: Signal Bool
           trIsNodeAlive =
                 not . Set.null
-            <$> Signal.keyedUntil (fromJoinedOrKilled (Set.singleton ())
-                                                      Set.empty)
-                                  (fromJoinedOrKilled Set.empty
-                                                      (Set.singleton ()))
+            <$> Signal.keyedUntil (fromJoinedOrTerminated (Set.singleton ())
+                                                           Set.empty)
+                                  (fromJoinedOrTerminated  Set.empty
+                                                          (Set.singleton ()))
                                   (const False)
                                   trJoinKillSig
 
@@ -2741,10 +3058,10 @@ prop_diffusion_target_active_below ioSimTrace traceNumber =
 
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
-              Signal.fromChangeEvents Killed -- Default to TrKillingNode
+              Signal.fromChangeEvents Terminated -- Default to TrKillingNode
             . Signal.selectEvents
                 (\case TrJoiningNetwork -> Just Joined
-                       TrKillingNode    -> Just Killed
+                       TrTerminated     -> Just Terminated
                        _                -> Nothing
                 )
             . selectDiffusionSimulationTrace
@@ -2759,10 +3076,10 @@ prop_diffusion_target_active_below ioSimTrace traceNumber =
           trIsNodeAlive :: Signal Bool
           trIsNodeAlive =
                 not . Set.null
-            <$> Signal.keyedUntil (fromJoinedOrKilled (Set.singleton ())
-                                                      Set.empty)
-                                  (fromJoinedOrKilled Set.empty
-                                                      (Set.singleton ()))
+            <$> Signal.keyedUntil (fromJoinedOrTerminated (Set.singleton ())
+                                                           Set.empty)
+                                  (fromJoinedOrTerminated  Set.empty
+                                                          (Set.singleton ()))
                                   (const False)
                                   trJoinKillSig
 
@@ -2891,14 +3208,14 @@ prop_diffusion_target_active_local_below ioSimTrace traceNumber =
 
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
-              Signal.fromChangeEvents Killed -- Default to TrKillingNode
+              Signal.fromChangeEvents Terminated -- Default to TrKillingNode
             . Signal.selectEvents
-                    (\case DiffusionDiffusionSimulationTrace TrJoiningNetwork
+                    (\case DiffusionSimulationTrace TrJoiningNetwork
                             -> Just Joined
-                           DiffusionDiffusionSimulationTrace TrKillingNode
-                            -> Just Killed
+                           DiffusionSimulationTrace TrTerminated
+                            -> Just Terminated
                            DiffusionConnectionManagerTrace CM.TrShutdown
-                            -> Just Killed
+                            -> Just Terminated
                            _ -> Nothing
                     )
             $ events
@@ -2912,10 +3229,10 @@ prop_diffusion_target_active_local_below ioSimTrace traceNumber =
           trIsNodeAlive :: Signal Bool
           trIsNodeAlive =
                 not . Set.null
-            <$> Signal.keyedUntil (fromJoinedOrKilled (Set.singleton ())
-                                                      Set.empty)
-                                  (fromJoinedOrKilled Set.empty
-                                                      (Set.singleton ()))
+            <$> Signal.keyedUntil (fromJoinedOrTerminated (Set.singleton ())
+                                                           Set.empty)
+                                  (fromJoinedOrTerminated  Set.empty
+                                                          (Set.singleton ()))
                                   (const False)
                                   trJoinKillSig
 
@@ -3058,7 +3375,8 @@ async_demotion_network_script =
 
     simArgs = SimArgs {
         saSlot             = secondsToDiffTime 1,
-        saQuota            = 5  -- 5% chance of producing a block
+        saQuota            = 5,  -- 5% chance of producing a block
+        saTxDecisionPolicy = defaultTxDecisionPolicy
       }
     peerTargets = Governor.nullPeerSelectionTargets {
       targetNumberOfKnownPeers = 1,
@@ -3084,7 +3402,8 @@ async_demotion_network_script =
         naChainSyncEarlyExit
                            = False,
         naPeerSharing      = PeerSharingDisabled,
-        naFetchModeScript  = singletonScript FetchModeDeadline
+        naFetchModeScript  = singletonScript (PraosFetchMode FetchModeDeadline),
+        naTxs              = []
       }
 
 
@@ -3324,10 +3643,10 @@ prop_diffusion_target_active_local_above ioSimTrace traceNumber =
 
           trJoinKillSig :: Signal JoinedOrKilled
           trJoinKillSig =
-              Signal.fromChangeEvents Killed -- Default to TrKillingNode
+              Signal.fromChangeEvents Terminated -- Default to TrKillingNode
             . Signal.selectEvents
                 (\case TrJoiningNetwork -> Just Joined
-                       TrKillingNode    -> Just Killed
+                       TrTerminated     -> Just Terminated
                        _                -> Nothing
                 )
             . selectDiffusionSimulationTrace
@@ -3342,10 +3661,10 @@ prop_diffusion_target_active_local_above ioSimTrace traceNumber =
           trIsNodeAlive :: Signal Bool
           trIsNodeAlive =
                 not . Set.null
-            <$> Signal.keyedUntil (fromJoinedOrKilled (Set.singleton ())
-                                                      Set.empty)
-                                  (fromJoinedOrKilled Set.empty
-                                                      (Set.singleton ()))
+            <$> Signal.keyedUntil (fromJoinedOrTerminated (Set.singleton ())
+                                                           Set.empty)
+                                  (fromJoinedOrTerminated  Set.empty
+                                                          (Set.singleton ()))
                                   (const False)
                                   trJoinKillSig
 
@@ -3645,7 +3964,7 @@ prop_unit_4258 =
                      abiSDUSize = LargeSDU
                    }
       diffScript = DiffusionScript
-        (SimArgs 1 10)
+        (SimArgs 1 10 defaultTxDecisionPolicy)
         (singletonTimedScript Map.empty)
         [( NodeArgs (-3) InitiatorAndResponderDiffusionMode (Just 224)
              Map.empty
@@ -3673,7 +3992,8 @@ prop_unit_4258 =
              (Script (DNSLookupDelay {getDNSLookupDelay = 0.065} :| []))
              Nothing
              False
-             (Script (FetchModeDeadline :| []))
+             (Script (PraosFetchMode FetchModeDeadline :| []))
+             []
          , [ JoinNetwork 4.166_666_666_666,
              Kill 0.3,
              JoinNetwork 1.517_857_142_857,
@@ -3715,7 +4035,8 @@ prop_unit_4258 =
                      ]))
              Nothing
              False
-             (Script (FetchModeDeadline :| []))
+             (Script (PraosFetchMode FetchModeDeadline :| []))
+             []
          , [ JoinNetwork 3.384_615_384_615,
              Reconfigure 3.583_333_333_333 [(1,1,Map.fromList [(RelayAccessAddress "0.0.0.4" 9,LocalRootConfig DoNotAdvertisePeer InitiatorAndResponderDiffusionMode IsNotTrustable)])],
              Kill 15.555_555_555_55,
@@ -3748,7 +4069,7 @@ prop_unit_reconnect :: Property
 prop_unit_reconnect =
   let diffScript =
         DiffusionScript
-          (SimArgs 1 10)
+          (SimArgs 1 10 defaultTxDecisionPolicy)
           (singletonTimedScript Map.empty)
           [(NodeArgs
               (-3)
@@ -3778,7 +4099,8 @@ prop_unit_reconnect =
               (Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []))
               Nothing
               False
-              (Script (FetchModeDeadline :| []))
+              (Script (PraosFetchMode FetchModeDeadline :| []))
+              []
           , [ JoinNetwork 0
             ])
           , (NodeArgs
@@ -3806,7 +4128,8 @@ prop_unit_reconnect =
              (Script (DNSLookupDelay {getDNSLookupDelay = 0} :| []))
              Nothing
              False
-             (Script (FetchModeDeadline :| []))
+             (Script (PraosFetchMode FetchModeDeadline :| []))
+             []
          , [ JoinNetwork 10
            ])
          ]
@@ -4221,12 +4544,13 @@ unit_peer_sharing =
         naDNSLookupDelayScript = singletonScript (DNSLookupDelay 0.01),
         naChainSyncEarlyExit = False,
         naChainSyncExitOnBlockNo = Nothing,
-        naFetchModeScript = singletonScript FetchModeDeadline,
-        naConsensusMode
+        naFetchModeScript = singletonScript (PraosFetchMode FetchModeDeadline),
+        naConsensusMode,
+        naTxs = []
       }
 
     script = DiffusionScript
-               (mainnetSimArgs 3)
+               (mainnetSimArgs 3 defaultTxDecisionPolicy)
                (singletonScript (mempty, ShortDelay))
                [ ( (defaultNodeArgs GenesisMode) { naAddr = ip_0,
                                      naLocalRootPeers = [(1, 1, Map.fromList [(ra_1, LocalRootConfig DoNotAdvertisePeer InitiatorAndResponderDiffusionMode IsNotTrustable)])],
@@ -4709,7 +5033,7 @@ unit_local_root_diffusion_mode diffusionMode =
 
     script =
       DiffusionScript
-        (SimArgs 1 20)
+        (SimArgs 1 20 defaultTxDecisionPolicy)
         (singletonTimedScript Map.empty)
         [ -- a relay node
           (NodeArgs {
@@ -4737,7 +5061,8 @@ unit_local_root_diffusion_mode diffusionMode =
              naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0.1} :| []),
              naChainSyncExitOnBlockNo = Nothing,
              naChainSyncEarlyExit = False,
-             naFetchModeScript = Script (FetchModeDeadline :| [])
+             naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+             naTxs = []
            }
           , [JoinNetwork 0]
           )
@@ -4771,7 +5096,8 @@ unit_local_root_diffusion_mode diffusionMode =
              naDNSLookupDelayScript = Script (DNSLookupDelay {getDNSLookupDelay = 0.1} :| []),
              naChainSyncExitOnBlockNo = Nothing,
              naChainSyncEarlyExit = False,
-             naFetchModeScript = Script (FetchModeDeadline :| [])
+             naFetchModeScript = Script (PraosFetchMode FetchModeDeadline :| []),
+             naTxs = []
            }
           , [JoinNetwork 0]
           )
@@ -4813,13 +5139,13 @@ prop_no_peershare_unwilling_iosim
 -- Utils
 --
 
-data JoinedOrKilled = Joined | Killed
+data JoinedOrKilled = Joined | Terminated
   deriving (Eq, Show)
 
 -- Similar to 'either' but for 'JoinedOrKilled'
-fromJoinedOrKilled :: c -> c -> JoinedOrKilled -> c
-fromJoinedOrKilled j _ Joined = j
-fromJoinedOrKilled _ k Killed = k
+fromJoinedOrTerminated :: c -> c -> JoinedOrKilled -> c
+fromJoinedOrTerminated j _ Joined     = j
+fromJoinedOrTerminated _ k Terminated = k
 
 getTime :: (Time, ThreadId (IOSim s), Maybe ThreadLabel, SimEventType) -> Time
 getTime (t, _, _, _) = t
@@ -4858,7 +5184,7 @@ selectDiffusionPeerSelectionEvents = Signal.selectEvents
 selectDiffusionSimulationTrace :: Events DiffusionTestTrace
                                -> Events DiffusionSimulationTrace
 selectDiffusionSimulationTrace = Signal.selectEvents
-                    (\case DiffusionDiffusionSimulationTrace e -> Just e
+                    (\case DiffusionSimulationTrace e -> Just e
                            _                                   -> Nothing)
 
 selectDiffusionPeerSelectionState :: Eq a
@@ -5003,7 +5329,7 @@ takeUntilEndofTurn n as =
 labelDiffusionScript :: DiffusionScript -> Property -> Property
 labelDiffusionScript (DiffusionScript args _ nodes) =
       label ("sim args: "
-              ++ show args)
+              ++ renderSimArgs args)
     . label ("Nº nodes: "
               ++ show (length nodes))
     . label ("Nº nodes in InitiatorOnlyDiffusionMode: "