+++ ouroboros-network/CHANGELOG.md
@@ -10,6 +10,7 @@
   are captured by the `DNSPeersKind` type, which also distinguishes the type
   of ledger peer.
 * Added `dispatchLookupWithTTL`
+* Added `Ouroboros.Network.TxSubmission.Inbound.V2`.
 
 ### Breaking changes
 
@@ -38,6 +39,10 @@
 - Renamed `Applications` to `DiffusionApplications`
 - `runM` function now receives `ExtraParameters` as an argument
 - Configurable Mux Egress Poll Interval
+- `Ouroboros.Network.TxSubmission.Inbound` moved to `Ouroboros.Network.TxSubmission.Inbound.V1`
+- `Ouroboros.Network.TxSubmission.Inbound.V1.txSubmissionInbound` takes extra argument: `TxSubmissionInitDelay` (previously configurable through `cabal` flags).
+- Removed the `txsubmission-delay` cabal flag.
+- `ProtocolErrorRequestedTooManyTxids` includes number of unacked txids.
 
 ## 0.21.1.0 -- 2025-05-26